requirements.txt
openai>=1.0.0
supabase>=2.0.0
streamlit>=1.30.0
python-dotenv
pandas
tiktoken
tqdm
pydantic-settings
pypdf>=3.0.0
google-generativeai>=0.8.0
tenacity
PyMuPDF>=1.23.0
playwright>=1.41.0
